In class detail, took out calculate_range_type() from DamageFunction Calculator
Just realized that the function calculate_range_type() is just extra code that does nothing to help reduce code. This is because, still need the if-else statmenets for the different calculations done. Having an extra if-else statement only adds more redundant code that does nothing.

Also to save execution time, decided to give DamageRangeFunction min and max range so have a ranges to compare to handy and not needing to call getter chain every time do calculations (which will be done a lot).