Updated Sequence diagram with change to how all hits to kill is calculated
Outside of updating the sequence diagram, also updated HitsToKillParser class as because it doesn't take in a damage_info reference, need to tell when parsing hits to kill to string if torso multi is 1.
That information changes the string as then instead of having torso hits seperate, say for example "2 torso/limb shots".