Update design diagram to better represent code
Found that the way code implements the design, the design lacked some things.

Two things added were in sequence diagram for calculating hits to kill, the entire loop to do so should be a self message as it is a function that is doing all those other function calls.

Other thing is activity diagram for calculating all combinations of hits to kill, the limb hits only hits to kill need to be slightly seperate as to tell the function if torso multi is 1 or not. Algorithm is the same, difference is the formatting of the string into the hits_to_kill dict when it comes to limb or having torso/limb.