Started testing calculate_all_combinations_hits_to_kill()
Did tests for TRG 42 and Gyrojet Mk 1. Still need to do test for TAR 21.
Found an issue in range calculations where super high decimal places can be off a bit. This is probably due to hardware issues of being unable to represent the decimal in binary.  To fix, decided to round to 3 decimal places on the return of the range.