Update weapon_labelframe.py so when radiobuttons are pressed, multi_frame gets unmapped or mapped
Added ability for the radio buttons to unmap the multiplier frame or map them back again, depending on option chosen.

Issue is that when unmapping, because using grid, actual placement changes. Seeing if can make the frame just invisible instead unmapping so there is no change in position of other widgets.