Finished testing of match_one_non_zero_num()
Also changed names of the function match_two_non_neg_nums() to match_two_nums().

This is due to how the function only searches for 2 numbers in general. It will still match a negative number (though ignores the '-' in front).

Appropriately, changed the name the the testsuite then as well to match the function it tests. Kept names of test same as data tested still matches name of tests.