Finished testing PF_regex.py functions
Finished testing the regex functions for now.
Won't be testing the verify functions as unsure how to do that with unittest when the class requires DamageFrame object which requires a tkinter.Misc object.
Also, the values within the DamageFrame which we extract and verify can't have the values affected outside from using the GUI.