Moved the command to the parent class of the ResultButton classes
This is due to the fact that damage_info_control makes a DamageInfo class, and the function that needs to be called is an abstract method that gets implemented by their child classes.

All in all, it allows for the child button classes to end up having to call the exact same function from the parent DamageInfo class, thus we only need to implement this command once, in the parent button class.