Implemented a pop-up window that prints out results for gun damage
At the moment the pop-up window only works for gun damage.
Still need to work on better results formatting, as the hits to kill aren't the best to read on user end.

This is due to the fact that grenade damage back-end has not been created yet, so pass on None to constructor of ResultButton, thus creating an error.

- FDIL