Made a table to show the results in
The table is finished, though not without bugs.

There are bugs with the label styles not being applied to Headers properly. Also now affecting main window labels.

Also, if there is too many results to go trough, it goes off page. Same deal with fullscreen if even that is not enough.