Made HitsToKillSorter
HitsToKillSorter is a data structure that converts the hits to kill from a dict to a sorted list, sorted by ascending order of range.
Overall this isn't an efficient algorithm, O(n log(n)). Not including python's insert() time complexity.
However, gets the job done and currently not thinking have to deal with large amounts of data per run.

Also implemented into the result window, so it currently works as intended.