Reformat hits to kill string
The hits to kill string has been reformatted with just numbers and |.
This makes it more CSV friendly.

This breaks the test_parse_tuple_to_string.py file (not including the fact it is already broken due to import issues).
What this means is the tests won't work as they aren't coded correctly as they expect old output, not the new one.