Changed constructors from result buttons, results table.
Changed the constructor of result table to instead of having a parameter of hits to kill, the parameter is now a damage info object.

This means change had to occur to result table constructor, thus how it was called in result_window.py

This was done so can have access to damage info when displaying the results.