result_tables has been reformatted
For 1, made the line number calculations of where to insert common between the classes by making the line_num a class attribute. This allows for adding new lines without having to know the exact line number at that time. (As long as rules to update it properly are followed).

Also added damage stats to top of the display, as part of the header.