Update damage_info_control.py
Added check for both damage being the same or not.
This is needed as the app ends up with a 0 division if the damage drop is 0.
Also this app won't support 0 damage drop as one can just simply use a calculator to see hits to kill as they are constant all ranges.