Add way to add hits to HitsToKill_Grouper
Added methods to add to hits to the dict. Also added a getter for access to the dict. No setter will be made.
Also changed hit_groups to be a SortedDict from sortedcontainers. Sorting by key is what is needed, so no need to define our own iter. Not yet tested if works.

Signed-off-by: FDIL <fdil.fadilh1@gmail.com>